% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{calculate_accuracy}
\alias{calculate_accuracy}
\title{Calculate Accuracy}
\usage{
calculate_accuracy(y_pred, y_test)
}
\arguments{
\item{y_pred}{A vector that contains the predicted labels.}

\item{y_test}{A vector that contains the true labels.}
}
\value{
A numeric value that represents the accuracy.
}
\description{
This function calculates the accuracy of the model's predictions.
}
