% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{calculate_similarities}
\alias{calculate_similarities}
\title{Calculate Similarities}
\usage{
calculate_similarities(
  model,
  words = list(c("her", "his"), c("her", "the")),
  vocab
)
}
\arguments{
\item{model}{A keras model object.}

\item{words}{A list of character vectors that represent word pairs. Default: list(c("her", "his"), c("her", "the"))}

\item{vocab}{A character vector that contains the vocabulary.}
}
\value{
A list that contains the cosine similarities for each word pair.
}
\description{
This function calculates the cosine similarity between word vectors.
}
