% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{create_table}
\alias{create_table}
\title{Create Table}
\usage{
create_table(results)
}
\arguments{
\item{results}{A data frame that contains the results of the simulations. It should have columns for "embedding_dim", "context_size", "train_accuracy", "mcse_train_accuracy", "test_accuracy", and "mcse_test_accuracy"}
}
\value{
A table that contains the selected columns.
}
\description{
This function selects specific columns from the results data frame and creates a table.
}
