% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model_prediction}
\alias{model_prediction}
\title{Model Prediction}
\usage{
model_prediction(model, x_test)
}
\arguments{
\item{model}{A keras model object.}

\item{x_test}{A data frame that contains the test data.}
}
\value{
A vector that contains the predicted labels.
}
\description{
This function used trained model to make predictions on the test set.
}
