% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model_training}
\alias{model_training}
\title{Model Training}
\usage{
model_training(
  x_train,
  y_train_onehot,
  x_val,
  y_val_onehot,
  vocab,
  embedding_dim = 30,
  context_size = 2,
  random_seed = 900,
  batch_size = 256,
  epochs = 10,
  h = 50,
  learning_rate = 0.005,
  early_stop_min_delta = 0.05,
  early_stop_patience = 3,
  verbose = 1
)
}
\arguments{
\item{x_train}{A data frame that contains the training data.}

\item{y_train_onehot}{A data frame that contains the one-hot encoded training labels.}

\item{x_val}{A data frame that contains the validation data.}

\item{y_val_onehot}{A data frame that contains the one-hot encoded validation labels.}

\item{vocab}{A character vector that contains the vocabulary.}

\item{embedding_dim}{An integer that specifies the dimension of the embedding layer. Default is 30.}

\item{context_size}{An integer that specifies the context size. Default is 2.}

\item{random_seed}{An integer that sets the seed for reproducibility. Default is 900.}

\item{batch_size}{An integer that specifies the batch size. Default is 256.}

\item{epochs}{An integer that specifies the number of epochs. Default is 10.}

\item{h}{An integer that specifies the number of units in the hidden layer. Default is 50.}

\item{learning_rate}{A numeric value that specifies the learning rate. Default is 5e-3.}

\item{early_stop_min_delta}{A numeric value that specifies the minimum change in the monitored quantity to qualify as an improvement. Default is 0.05.}

\item{early_stop_patience}{An integer that specifies the number of epochs with no improvement after which training will be stopped. Default is 3.}

\item{verbose}{An integer that specifies the verbosity mode. Default is 1.}
}
\value{
A list that includes the trained model and the final training accuracy.
}
\description{
This function creates and trains a neural network model.
}
