% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_test_accuracy}
\alias{plot_test_accuracy}
\title{Plot Test Accuracy with MCSE}
\usage{
plot_test_accuracy(
  results,
  param_name,
  param_values,
  colors = c("red", "blue", "yellow", "red", "blue"),
  linetypes = c("solid", "dashed", "solid", "dashed", "solid"),
  plot_title = "Test Accuracy with MCSE"
)
}
\arguments{
\item{results}{A data frame that contains the results of the simulations. It should have columns for 'param_value', 'test_accuracy', 'test_lower_bound', 'test_upper_bound', and 'embedding_dim'.}

\item{param_name}{A string that specifies the name of the parameter that was varied in the simulations.}

\item{param_values}{A numeric vector that contains the values of the parameter that were used in the simulations.}

\item{colors}{A character vector that specifies the colors to be used in the plot. Default is c("red", "blue", "yellow", "red", "blue").}

\item{linetypes}{A character vector that specifies the line types to be used in the plot. Default is c("solid", "dashed", "solid", "dashed", "solid").}

\item{plot_title}{A string that specifies the title of the plot. Default is "Test Accuracy with MCSE".}
}
\value{
A ggplot object that represents the plot.
}
\description{
This function plots the test accuracy with Monte Carlo Standard Error (MCSE) for different parameter values and embedding dimensions.
}
\examples{
# to draw the plot for embedding dimension of accuracy vs. context size
# plot_test_accuracy(results, param_name="context_size", param_values = c(3,4,5))
}
