% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{process_data}
\alias{process_data}
\title{Process Data}
\usage{
process_data(data_path, context_size = 3, lowest_frequency = 3)
}
\arguments{
\item{data_path}{A string that specifies the path to the data file.}

\item{context_size}{An integer that specifies the context size which is the length of the input word sequence to the model. Default is 3.}

\item{lowest_frequency}{An integer that specifies the lowest frequency for a word to be included in the vocabulary.
The word with frequency less than 3 will be replaced to a special token <UNK> . Default is 3.}
}
\value{
A list that includes the processed data (input word sequence and target subsequent word) and
       the vocabulary (a table for mapping a word with its unique word id).
}
\description{
This function loads the data, converts all sentences to lowercase, creates term frequency of the words, and converts words to unique ids.
}
