% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{run_simulation}
\alias{run_simulation}
\title{Run Simulation}
\usage{
run_simulation(
  n_simulations = 10,
  data_path = "rocstories.csv",
  random_seed = 900,
  context_size = 3,
  train_portion = 0.8,
  val_portion = 0.1,
  test_portion = 0.1,
  lowest_frequency = 3,
  embedding_dim = 30,
  batch_size = 256,
  epochs = 10,
  h = 50,
  learning_rate = 0.005,
  early_stop_min_delta = 0.05,
  early_stop_patience = 3,
  verbose = 1
)
}
\arguments{
\item{n_simulations}{An integer that specifies the number of simulations to run. Default is 10.}

\item{data_path}{A string that specifies the path to the data file. Default is "rocstories.csv".}

\item{random_seed}{An integer that sets the seed for reproducibility. Default is 900.}

\item{context_size}{An integer that specifies the context size. Default is 3.}

\item{train_portion}{A numeric value that specifies the portion of the data to be used for training. Default is 0.8.}

\item{val_portion}{A numeric value that specifies the portion of the data to be used for validation. Default is 0.1.}

\item{test_portion}{A numeric value that specifies the portion of the data to be used for testing. Default is 0.1.}

\item{lowest_frequency}{An integer that specifies the lowest frequency. Default is 3.}

\item{embedding_dim}{An integer that specifies the dimension of the embedding layer. Default is 30.}

\item{batch_size}{An integer that specifies the batch size. Default is 256.}

\item{epochs}{An integer that specifies the number of epochs. Default is 10.}

\item{h}{An integer that specifies the number of units in the hidden layer. Default is 50.}

\item{learning_rate}{A numeric value that specifies the learning rate. Default is 5e-3.}

\item{early_stop_min_delta}{A numeric value that specifies the minimum change in the monitored quantity to qualify as an improvement. Default is 0.05.}

\item{early_stop_patience}{An integer that specifies the number of epochs with no improvement after which training will be stopped. Default is 3.}

\item{verbose}{An integer that specifies the verbosity mode. Default is 1.}
}
\value{
A list that includes the mean test accuracy, the Monte Carlo Standard Error of the test accuracy, the mean train accuracy, and the Monte Carlo Standard Error of the train accuracy.
}
\description{
This function runs a series of simulations for the neural network model.
}
\examples{
# run simulation with default hyperparameter settings
run_simulation(n_simulations=10, data_path="rocstories.csv", random_seed=900)

}
