% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{run_simulations}
\alias{run_simulations}
\title{Run Simulations}
\usage{
run_simulations(embedding_dim_values, param_values, param_name, other_params)
}
\arguments{
\item{embedding_dim_values}{A numeric vector that contains the values of the embedding dimension to be tested.}

\item{param_values}{A numeric vector that contains the values of the target hyperparameter to be tested.}

\item{param_name}{A character string that specifies the name of the target hyperparameter. Could be one of "context_size", "h", or "batch_size". Default is "context_size".}

\item{other_params}{A list that contains the values of the other hyperparameters. The list should include the following elements:
* data: A data frame that contains the data to be processed.
* n_simulations: An integer that specifies the number of simulations to run. Default is 50.
* random_seed: An integer that sets the seed for reproducibility. Default is 900.
* train_portion: A numeric value that specifies the portion of the data to be used for training. Default is 0.8.
* val_portion: A numeric value that specifies the portion of the data to be used for validation. Default is 0.1.
* test_portion: A numeric value that specifies the portion of the data to be used for testing. Default is 0.1.
* lowest_frequency: An integer that specifies the lowest frequency of words to be included in the vocabulary. Default is 3.
* epochs: An integer that specifies the number of epochs. Default is 20.
* learning_rate: A numeric value that specifies the learning rate. Default is 5e-3.
* early_stop_min_delta: A numeric value that specifies the minimum change in the monitored quantity to qualify as an improvement. Default is 0.01.
* early_stop_patience: An integer that specifies the number of epochs with no improvement after which training will be stopped. Default is 2.
* verbose: An integer that specifies the verbosity mode. Default is 1.}
}
\value{
A data frame that includes the results of the simulations.
}
\description{
This function runs simulations for combinations of embedding size and target hyperparameters.
}
\examples{
# define hyperparameters you want to explore
embedding_dim_values <- c(10000,3319,300,30,3)
context_size_values <- c(2, 3, 4)
load(rocstories)
# must do: define the parameters passed to the run_simulation function
other_params <- list(data = rocstories, n_simulations = 50, random_seed = 900, train_portion = 0.8, val_portion = 0.1, test_portion = 0.1, lowest_frequency = 3,epochs=20, learning_rate=5e-3, early_stop_min_delta=0.01,early_stop_patience=2,verbose=1)
results <- run_simulations(embedding_dim_values=embedding_dim_values, param_values=context_size_values, param_name="context_size", other_params)


}
