% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{split_data}
\alias{split_data}
\title{Split Data}
\usage{
split_data(
  x_data,
  y_data,
  vocab,
  random_seed = 900,
  train_portion = 0.8,
  val_portion = 0.1,
  test_portion = 0.1
)
}
\arguments{
\item{x_data}{A data frame that contains the processed data.}

\item{y_data}{A data frame that contains the labels.}

\item{vocab}{A table for mapping a word with its unique word id.}

\item{random_seed}{An integer that sets the seed for reproducibility. Default is 900.}

\item{train_portion}{A numeric value between 0 and 1 that specifies the portion of the data to be used for training. Default is 0.8.}

\item{val_portion}{A numeric value between 0 and 1 that specifies the portion of the data to be used for validation. Default is 0.1.}

\item{test_portion}{A numeric value between 0 and 1 that specifies the portion of the data to be used for testing. Default is 0.1.}
}
\value{
A list that includes the training, validation, and test sets, and the vocabulary (a table for mapping a word with its unique word id).
}
\description{
This function shuffles and splits the data into training, validation, and test sets.
}
